Added questions page
Changed "Write a post" button to "Ask a question", with a simple form that submits a Github issue to the knowledge repo.

The knowledge repo URI and user are hard-coded for now. Need to extract from the GitRepository object, and handle user identity and auth (use Github creds or separate user databse for the Knowledge Repo?)